cam tests failed.
rigidbody is incompatible with 3rd person camera sample. might be possible with custom solution. Requires creating an aiming core (child of player) that rotates in the direction of the mouse movement. rewires the input such that forwards is where the ball is pointing. Alternative is to give the player the ability to switch cameras from front to back (easier)